content_creation-mcp/requirements.txt
fastapi
pydantic 
uvicorn
streamlit
pyyaml
langchain
langchain-openai
python-dotenv
fastmcp